Move Proxy type declarations into a header, add to_string for MutableProxy
